+++ crates/store/src/db/models/queries/accounts.rs
@@ -50,8 +50,7 @@ use crate::errors::DatabaseError;
 
 mod at_block;
 pub(crate) use at_block::{
-    select_account_header_at_block,
-    select_account_storage_at_block,
+    select_account_header_with_storage_header_at_block,
     select_account_vault_at_block,
 };